+++ template/src/bin/async_main.rs
@@ -77,7 +77,7 @@ async fn main(spawner: Spawner) -> ! {
 
     //IF option("alloc")
     //REPLACE 65536 max-dram2-uninit
-    esp_alloc::heap_allocator!(#[unsafe(link_section = ".dram2_uninit")] size: 65536);
+    esp_alloc::heap_allocator!(#[esp_hal::ram(reclaimed)] size: 65536);
     //IF option("wifi") && (option("ble-bleps") || option("ble-trouble"))
     // COEX needs more RAM - so we've added some more
     esp_alloc::heap_allocator!(size: 64 * 1024);
@@ -88,9 +88,10 @@ async fn main(spawner: Spawner) -> ! {
     //IF option("esp32") || option("esp32s2") || option("esp32s3")
     esp_rtos::start(timg0.timer0);
     //ELSE
-    let sw_interrupt = esp_hal::interrupt::software::SoftwareInterruptControl::new(peripherals.SW_INTERRUPT);
+    let sw_interrupt =
+        esp_hal::interrupt::software::SoftwareInterruptControl::new(peripherals.SW_INTERRUPT);
     esp_rtos::start(timg0.timer0, sw_interrupt.software_interrupt0);
-    //ENDIF 
+    //ENDIF
 
     //IF option("defmt") || option("log")
     info!("Embassy initialized!");
@@ -99,12 +100,12 @@ async fn main(spawner: Spawner) -> ! {
     //ENDIF
 
     //IF option("ble-trouble") || option("ble-bleps") || option("wifi")
-    let radio_init = esp_radio::init()
-    .expect("Failed to initialize Wi-Fi/BLE controller");
+    let radio_init = esp_radio::init().expect("Failed to initialize Wi-Fi/BLE controller");
     //ENDIF
     //IF option("wifi")
-    let (mut _wifi_controller, _interfaces) = esp_radio::wifi::new(&radio_init, peripherals.WIFI, Default::default())
-        .expect("Failed to initialize Wi-Fi controller");
+    let (mut _wifi_controller, _interfaces) =
+        esp_radio::wifi::new(&radio_init, peripherals.WIFI, Default::default())
+            .expect("Failed to initialize Wi-Fi controller");
     //ENDIF
     //IF option("ble-trouble")
     // find more examples https://github.com/embassy-rs/trouble/tree/main/examples/esp32